Simplify connections
